# docs/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "This page is for developers. We built this site entirely off the APIs below and we want to share them with you.",
		"version": "0.0.8",
		"title": "API Documentation for Developers"
	},
	"tags": [
		{
			"name": "  ",
			"description": "Forest Service ePermit API"
		}
	],
	"securityDefinitions": {
		 "access_token": {
			"type": "apiKey",
			"name": "accessToken",
			"in": "header"
		 }
	 },
	"paths": {
		"/auth": {
			"post": {
				"tags": [
					"Authorization Request"
				],
				"summary": "Authorization Request",
				"description": "",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						 "name": "username",
						 "in": "formData",
						 "description": "Username",
						 "required": true,
						 "type": "string"
					},
					{

						 "name": "password",
						 "in": "formData",
						 "description": "Password",
						 "required": true,
						 "type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
		"/permits/applications/special-uses/noncommercial/": {
			"post": {
				"tags": [
					"Noncommercial Permit"
				],
				"summary": "Submit a new noncommercial permit",
				"description": "",
				"operationId": "createPermit",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"description": "noncommercial permit information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/noncommercialPermit"
						}
					},
					{
						"name": "x-access-token",
						"in": "header",
						"description": "access token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Invalid input"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
		"/permits/applications/special-uses/noncommercial/{controlNumber}": {
			"put": {
				"tags": [
					"Noncommercial Permit"
				],
				"summary": "Update an existing noncommercial permit",
				"description": "",
				"operationId": "updatePermit",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					 {
						"name": "controlNumber",
						"in": "path",
						"description": "ID of noncommercial permit to be updated",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Permit object that needs to be updated",
						"required": true,
						"schema": {
							"$ref": "#/definitions/noncommercialPermit"
						}
					},
					{
						"name": "x-access-token",
						"in": "header",
						"description": "access token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Invalid ID supplied"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Permit not found"
					},
					"405": {
						"description": "Validation exception"
					}
				}
			},
			"get": {
				"tags": [
					"Noncommercial Permit"
				],
				"summary": "Find noncommercial permit by ID",
				"description": "Returns a single permit",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "controlNumber",
						"in": "path",
						"description": "ID of noncommercial permit to return",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"name": "x-access-token",
						"in": "header",
						"description": "access token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/noncommercialPermitGet"
						}
					},
					"400": {
						"description": "Invalid ID supplied"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Permit not found"
					}
				}
			}
		},
		"/permits/applications/special-uses/commercial/temp-outfitters/": {
			"post": {
				"tags": [
					"Outfitter and Guide Permit"
				],
				"summary": "Submit a new temp-outfitter permit",
				"description": "",
				"consumes": [
					"multipart/form-data"
				],
				"produces": [
					"application/json"
				],
				"parameters": [					
					{
						"in": "formData",
						"name": "body",
						"description": "temp-outfitter permit information",
						"required": true,
						"schema": {
							"$ref": "#/definitions/tempOutfitterPermit"
						}
					},
					{
						"in": "formData",
						"name": "guideDocumentation",
						"description": "temp-outfitter guideDocumentation file upload",
						"required": true,
						"type": "file"
					},
					{
						"in": "formData",
						"name": "acknowledgementOfRiskForm",
						"description": "temp-outfitter acknowledgementOfRiskForm file upload",
						"required": true,
						"type": "file"
					},
					{
						"in": "formData",
						"name": "insuranceCertificate",
						"description": "temp-outfitter insuranceCertificate file upload",
						"required": true,
						"type": "file"
					},
					{
						"in": "formData",
						"name": "goodStandingEvidence",
						"description": "temp-outfitter goodStandingEvidence file upload",
						"required": true,
						"type": "file"
					},
					{
						"in": "formData",
						"name": "operatingPlan",
						"description": "temp-outfitter operatingPlan file upload",
						"required": true,
						"type": "file"
					},
					{
						"name": "x-access-token",
						"in": "header",
						"description": "access token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Invalid input"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
		"/permits/applications/special-uses/commercial/temp-outfitters/{controlNumber}": {
			"put": {
				"tags": [
					"Outfitter and Guide Permit"
				],
				"summary": "Update an existing temp-outfitter Permit",
				"description": "",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					 {
						"name": "controlNumber",
						"in": "path",
						"description": "ID of temp-outfitter permit to be updated",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"in": "body",
						"name": "body",
						"description": "Permit object that needs to be updated",
						"required": true,
						"schema": {
							"$ref": "#/definitions/tempOutfitterPermit"
						}
					},
					{
						"name": "x-access-token",
						"in": "header",
						"description": "access token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "Success"
					},
					"400": {
						"description": "Invalid ID supplied"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Permit not found"
					},
					"405": {
						"description": "Validation exception"
					}
				}
			},
			"get": {
				"tags": [
					"Outfitter and Guide Permit"
				],
				"summary": "Find temp-outfitter permit by ID",
				"description": "Returns a single permit",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "controlNumber",
						"in": "path",
						"description": "ID of temp-outfitter permit to return",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"name": "x-access-token",
						"in": "header",
						"description": "access token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "successful operation",
						"schema": {
							"$ref": "#/definitions/tempOutfitterPermitGet"
						}
					},
					"400": {
						"description": "Invalid ID supplied"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Permit not found"
					}
				}
			}
		}
	},
	"definitions": {
		  "applicantInfoNoncommercial": {
			"type": "object",
			"properties": {
				"firstName": { "type": "string" },
				"lastName": { "type": "string" },
				"dayPhone": { "$ref": "#/definitions/phoneNumber" },
				"eveningPhone": { "$ref": "#/definitions/phoneNumber" },
				"emailAddress": { "type": "string" },
				"mailingAddress": { "type": "string" },
				"mailingAddress2": { "type": "string" },
				"mailingCity": { "type": "string" },
				"mailingState": { "type": "string" },
				"mailingZIP": { "type": "integer" },
				"organizationName": { "type": "string"},
				"website":{"type": "string"},
				"orgType":{
					"type":"string",
					"description":"Organization Type",
					"enum":[
						"Individual",
						"Corporation",
						"Limited Liability Company",
						"Partnership or Association",
						"State Government or Agency",
						"Local Government or Agency",
						"Nonprofit"
					]
				}				
			},
			"required": ["firstName","lastName","dayPhone","emailAddress","mailingAddress","mailingCity","mailingZIP","mailingState"],
			"xml": {
				"name": "applicantInfoNoncommercial"
			}
		},
		"applicantInfoTempOutfitter": {
			"type": "object",
			"properties": {
				"firstName": { "type": "string" },
				"lastName": { "type": "string" },
				"dayPhone": { "$ref": "#/definitions/phoneNumber" },
				"eveningPhone": { "$ref": "#/definitions/phoneNumber" },
				"emailAddress": { "type": "string" },
				"mailingAddress": { "type": "string" },
				"mailingAddress2": { "type": "string" },
				"mailingCity": { "type": "string" },
				"mailingState": { "type": "string" },
				"mailingZIP": { "type": "integer" },
				"organizationName": { "type": "string"},
				"website":{"type": "string"},
				"orgType":{
					"type":"string",
					"description":"Organization Type",
					"enum":[
						"Individual",
						"Corporation",
						"Limited Liability Company",
						"Partnership or Association",
						"State Government or Agency",
						"Local Government or Agency",
						"Nonprofit"
					]
				}				
			},
			"required": ["firstName","lastName","dayPhone","emailAddress","mailingAddress","mailingCity","mailingZIP","mailingState","orgType"],
			"xml": {
				"name": "applicantInfoTempOutfitter"
			}
		},
		"phoneNumber": {
		  "type": "object",
			"properties": {
				"areaCode": { "type": "integer"},
				"number": { "type": "integer"},
				"extension": { "type": "integer"},
				"type": {"type": "string"}
			},
			"required": ["areaCode","number","type"],
			"xml": {
				"name": "phoneNumber"
			}  
		},
		"noncommercialFields": {
		  "type": "object",
		  "properties": {
			 "activityDescription": { "type": "string"},
			 "locationDescription": {  "type": "string"},
			 "startDateTime": { "type": "string"},
			 "endDateTime": { "type": "string"},
			 "numberParticipants" : { "type": "integer"}
		  },
		  "required": ["activityDescription","locationDescription","startDateTime","endDateTime","numberParticipants"],
		  "xml": {
				"name": "noncommercialFields"
			} 
		},
		"tempOutfitterFields": {
		  "type": "object",
		  "properties": {
			 "activityDescription": { "type": "string"},
			 "locationDescription": {  "type": "string"},
			 "startDateTime": { "type": "string"},
			 "endDateTime": { "type": "string"},
			 "insuranceCertificate": { "type": "string"},
			 "goodStandingEvidence": {
				"description": "Certificate of good standing, operating agreement, or other evidence the business is in good standing",
				"type": "string"
			 },
			 "operatingPlan": {
				"description": "A completed Word or RTF operating plan template.",
				"type": "string"
			 }
		  },
		  "required": ["insuranceCertificate"],
		  "xml": {
				"name": "tempOutfitterFields"
			} 
		},
		"noncommercialPermit": {
			"type": "object",
			"properties": {
				"region": { "type" : "integer" },
				"forest": { "type" : "integer" },
				"district": { "type" : "integer" },
				"authorizingOfficerName": { "type" : "string"},
				"authorizingOfficerTitle": { "type" : "string"},
				"applicantInfo": {
					"$ref": "#/definitions/applicantInfoNoncommercial"
				},
				"type": {
					"type": "string",
					"enum":[
						"noncommercial",
						"tempOutfitterGuide"
					],
					"default": "noncommercial"
				},
				"noncommercialFields": {
				  "$ref": "#/definitions/noncommercialFields"
				}
			},
			"required": ["region","forest","district","applicantInfo","type","noncommercialFields"],
			"xml": {
				"name": "noncommercialPermit"
			}
		},
		"noncommercialPermitGet": {
			"type": "object",
			"properties": {
				"controlNumber": {"type": "integer"},
				"region": { "type" : "integer" },
				"forest": { "type" : "integer" },
				"district": { "type" : "integer" },
				"authorizingOfficerName": { "type" : "string"},
				"authorizingOfficerTitle": { "type" : "string"},
				"applicantInfo": {
					"$ref": "#/definitions/applicantInfoNoncommercial"
				},
				"type": {
					"type": "string", 
					"default": "noncommercial",
					"enum":[
						"noncommercial"
					]
				},
				"noncommercialFields": {
				  "$ref": "#/definitions/noncommercialFields"
				}
			},
			"required": ["controlNumber","region","forest","district","applicantInfo","type","noncommercialFields"],
			"xml": {
				"name": "noncommercialPermitGet"
			}
		},
		"tempOutfitterPermit": {
			"type": "object",
			 "properties": {
				"region": { "type" : "integer" },
				"forest": { "type" : "integer" },
				"district": { "type" : "integer" },
				"authorizingOfficerName": { "type" : "string"},
				"authorizingOfficerTitle": { "type" : "string"},
				"applicantInfo": {
					"$ref": "#/definitions/applicantInfoTempOutfitter"
				},
				"type": {
					"type": "string", 
					"enum":[
						"tempOutfitterGuide",
						"noncommercial"
					],
					"default": "tempOutfitterGuide"
				},
				"tempOutfitterFields": {
				  "$ref": "#/definitions/tempOutfitterFields"
				}
			 },
			 "required": ["region","forest","district","applicantInfo","type","tempOutfitterFields"],
			"xml": {
				"name": "tempOutfitterPermit"
			}
		},
		"tempOutfitterPermitGet": {
			"type": "object",
			 "properties": {
				"controlNumber": { "type": "integer"},
				"region": { "type" : "integer" },
				"forest": { "type" : "integer" },
				"district": { "type" : "integer" },
				"authorizingOfficerName": { "type" : "string"},
				"authorizingOfficerTitle": { "type" : "string"},
				"applicantInfo": {
					"$ref": "#/definitions/applicantInfoTempOutfitter"
				},
				"type": {
					"type": "string", 
					"default": "tempOutfitterGuide",
					"enum":[						
						"tempOutfitterGuide"
					]
				},
				"tempOutfitterFields": {
				  "$ref": "#/definitions/tempOutfitterFields"
				}
			 },
			 "required": ["controlNumber","region","forest","district","applicantInfo","type","tempOutfitterFields"],
			"xml": {
				"name": "tempOutfitterPermitGet"
			}
		}
	}
}